Multiple cameras | step 9
This change creates the notion of lead and follower schedules, aiming to build multiple camera schedules into the scheduler. The newly created follower schedules are those that run in parallel with their corresponding lead schedule.

A follower schedule inherits several attributes from its lead: target including alignment, start time and constraints, while the capture sequence and the stop conditions are uniquely might differ from the lead.

In this step, there is no synchronization of actions across lead and follower. The lead controls for example dithering or meridian flip and does not consider what the follower does. This will be the issue of one of the next steps.

Scheduling now consist of these steps:
1. Evaluate all jobs (including the follower jobs) and update whether they could be scheduled or have been completed.
2. Create a schedule for lead jobs only and select the lead job that should be executed next.
3. Start the lead job and all its follower jobs that aren't complete.
4. If the execution switches to a new lead job, all (not completed) follower jobs of the previous lead job will be aborted.