Add option to always invent the guidestar in MultiStar guiding.
Add an "experimental" option to always invent the guidestar in MultiStar guiding, hopefully making it robust to double stars. 

See https://indilib.org/forum/ekos/14852-weird-failure-with-internal-guider.html
There have been issues where a guide-star that's near another star causes issues with guiding. This may be due to inconsistent star detection, but in any case, the many reference stars in the MultiStar algorithm should be used to anchor the guiding instead of one guide star.

This MR also increases the default for max number of reference stars to 50, from 10.