Multiple cameras | step 10
Handling focusing for multiple optical trains on a single mount is the next step towards full multi camera support. Similar to the already implemented approach for capture module, the focus module obtains tabs where each tab covers focusing for one of the optical trains.

With this approach it is possible running multiple focus procedures in parallel for each optical train.

Events between the capture module and the focus module know their optical train making it possible to route the events from capture to focus on the same optical train back and forth.