Bug fix in scheduler for interface changes of focus events
With the introduction of multi focusers, the event interfaces have changed. This needs to be considered in the scheduler as well. This fixes the problem that the scheduler hangs after the focusing step.

This is only a bug fix. Full support of the multi focuser option in the scheduler will come later.