Multiple cameras | step 11 | dither sync
Dithering while capturing with multiple cameras leads to corrupted images when dithering happens while capturing on the other trains is running. Therefore a synchronization between dithering and capturing is required.

As soon as one camera requests dithering, this request is registered, but not executed immediately. Instead, all other cameras that are capturing currently are considered:
- If the remaining exposure time of a running camera is less than 50% of the exposure time from the camera, that requested the sync, pausing the capture process of that camera is requested.
- In the other case, capturing of the camera is suspended, since the camera requesting the dithering would have to wait for too long.
- As soon as all cameras have either been paused or suspended, dithering is started.

A completed dithering (may it be successful or not) will trigger all cameras to continue capturing.