Multiple cameras | step 14 | Meridian flip
This is a basic implementation to handle meridian flips with multiple cameras. The main camera keeps the control when a meridian flip should be started.

As soon as a meridian flip starts, the follower cameras are suspended and will be restarted as soon as the flip has completed and - if guiding was on - guiding has restarted.