Ignore device events if preparation has been completed
Restrict device events in SequenceJobState to the preparation state. Otherwise, this could cause ugly reentrances of capture starts.

Without this change, capture does not wait for a user requested focus run to be terminated disturbing the focus run. To reproduce it, you need to capture with a filter that has another filter set to be used for focusing.