Remove the "Simulate Eyepiece View" feature and strip down `EyepieceField`
The idea here is that the offerings of the eyepiece view feature have
already been superceded by two more powerful and easier-to-use
features in KStars:

1. The HiPS Overlay
2. The "Views" feature

The only feature provided by the old Eyepiece View dialog that is
perhaps removed without alternative is the ability to print a chart
with ticks indicating the orientation of the chart at different times
in a Dobsonian eyepiece. This feature was crude, likely buggy,
unmaintained and not universal. It was probably never used outside of
my testing. Similar results can instead be achieved using a DBus-based
Python script, if ever desired, with new DBus methods to be added in a
different commit.

As a result, this commit:

+ Removes the "Simulate Eyepiece View" feature from the popup menus –
  on the sky-map and in Observation Planner, and the corresponding
  dialog UI

+ Retains the DBus method for backwards compatibility