Capture Video Sequences
Capture sequences may be used now to capture video sequences.

Fortunately, the capture sequence format is already capable for the definition of videos. The "repeat" counter from still frames capturing is used for the definition of number of frames a video should last. Therefore, the only syntactical change was to introduce CAPTURE_TYPE_VIDEO as new frame type.