Enable Focuser Controls when Camera Disconnected
Currently focuser controls in the Focus tab are greyed out along with camera controls when the camera is disconnected.

This MR leaves the focuser controls (move to, move in, move out) enabled when the camera is disconnected.

The purpose of this is to facilitate the use-case where KStars / Ekos is used to initially focus but then another application, e.g. FireCapture is used to capture images. In this case allowing the user to manually tweak focus using Ekos controls is beneficial.

Reconnecting the camera will re-enable all controls and functionality.