Allow framing on all focus measures
Framing currently works for a measure=HFR. This MR will allow framing on any of the available measures as selected in Options.

The purpose of this change, other than to allow functionality available in Autofocus to be used when framing, is to allow users without an automatic focuser to make use of any of the available measures for manual focusing.