Reuse calibration after rotation: Part 1
Analyzing the possibility to reuse calibration after a rotation (*) of the guide cam results in the presented ~~draft~~ commit. In order to relate the calibration rotation with the camera position angle it is necessary to consolidate their angle measurements. The following changes were necessary to achieve this goal:
- some minor modifications in naming of variables for better
  understanding
- consolidation in rotation and angle senses (especially RA & DEC)
- consolidation in direction of increments/decrements of RA/DEC
- removal of several loose ad-hoc negations of coordinates
- some tuning of information in calibration plot
- correction of E-W drift and pulse direction in drift graph
- draw RADEC coordinate system in calibration plot
- adjust tests to new angle measurement

I hope there are no interferences with PHD2 or polar alignment. Reviewer and testers are welcome! 

**N.B. To operate this commit with the telescope simulator it is imperative to deploy the corresponding commit for the INDI telescope simulator driver!** (https://github.com/indilib/indi/pull/2153)

(*) This was a wish of a user in the forum some time ago. Alas I cannot find my contributions under "My Topics" anymore and so I couldn't find this record anymore.\
![Calibration](/uploads/7116c0b8eeb6414bfb23e1e852cedc23/Calibration.gif)\
Guiding in simulators after a manual pulse to the west and then to the north with camera PA -90°