DBus communication with FocusModule corrected
When introducing multiple focuser handling, the DBus interface remained on the level of single focusers. With this change, the FocusModule is the new DBus interface and each function now contains the optical train as argument to address the desired focuser.

With this fix it was possible to use autofocus runs on all optical trains in parallel inside the scheduler.