Focus speedup with shift and ctrl
Holding the shift or the ctrl key speeds up the manual motions of the focuser. Both shift and ctrl have a configurable factor which is applied to the standard (initial) step size for manual focus motions. If both shift and ctrl are pressed, their product is used as speed up factor.