Autofocus Optimisation
The Autofocus Optimisation (AO) change is intended to allow users to set a configurable timezone that allows an Autofocus request to be "optimised out" whereby the results of the previous Autofocus run are used, rather than running a new Autofocus.

The benefit of this is that, because several modules within Ekos can trigger an Autofocus in certain configurations, Autofocus can be repeated with little benefit to the user. By optimising out an Autofocus request the time taken to run Autofocus is saved.

AO will work only with Linear 1 Pass and Linear focus algorithms.

AO is enabled for automated requests from the Scheduler and Capture. Manual requests, Build Filter Offsets and Autofocus after a Meridian Flip will never be optimised out.

AO allows a user configurable parameter to be set...
![Screenshot_2025-01-31_at_21.36.20](/uploads/d2585e7fd7e22938fc2049a028c5c313/Screenshot_2025-01-31_at_21.36.20.png) 

If AF Optimise is set to 0s (the default) this feature is deactivated.

Setting AF Optimise to, say, 300s sets up a timezone of 5 mins. When an Autofocus request is received, AO checks whether a successful Autofocus completed within timezone (5mins) and if so, the Autofocus is optimised out.

AO will honour focus offsets.

AO with honour Adapt Start Pos (if this option is set) by adjusting the focuser position for temperature / altitude changes consistent with Adaptive Focus.