Fix Scheduler continues after Focus Fail
If a schedule is setup with a filter requiring an Autofocus, then if the Autofocus fails (Focus tries 3 times) then the schedule continues with Capture, effectively ignoring the Autofocus fail.

The reason the bug exists is that Capture attempts to set the filter which works, but if the associated Autofocus fails, then when the Scheduler reruns the job it finds that the correct filter is set and assumes that is good enough to proceed.

The fix in FilterManager maintains a "last Autofocus failed" flag per filter which is used if the to trigger a retry Autofocus if it previously fails which is consistent with the treatment of other Autofocus fails, e.g. a time based Autofocus.

If the sequence is being run from Capture then there is no change to behaviour... Capture will abort the sequence.