Init weather device selection
While beta testing my changes to the indi_ocs driver a it was found that if multiple weather devices are available at launch of Ekos the initial chosen device in the weatherSourceCombo is not used to filter the presented measurement until the user manually changes the selection. This adds a call to setWeatherSource from within initWeather.