Fixing a DBus issue and 2 Qt6 issues
Fixing a DBus loading issue encountered on MacOS when upgrading to Sequoia 15.4 the other day. Also fixing a QT6 issue with the Imaging Planner because KMessageBox Yes-No no longer exists, Warning-Continue-Cancel is now preferred. Finally fixing a crash issue with QT6 and Qt Unique Connection. Apparently that only is allowed with member functions, not with lambdas. The compiler says it is a warning, but when you run the program it causes a crash.

Please check to make sure these changes don't cause any undesired side effects.