Added pre & post driver shutdown scripts
While thinking about a user raised request in thread https://indilib.org/forum/development/15354-example-pre-post-job-script.html I realised that the requirement would be best met by having more options of when a user script could be triggered. This MR add the options for a pre & post driver shutdown script to go with the existing pre & post driver startup scripts.

I also added some labelling in the script selection dialog and fixed a minor bug in there as well (not limiting to driver count as an empty remote was still passed).