Show guiding image for PHD2
Display the guiding frame of PHD2 inside of the guiding module, similar to the internal guider. For devices with limited resources, this feature can be turned off with a newly introduced button.

Unfortunately, PHD2 reports only its lock position and not all stars that are considered for guiding. Therefore, even for multistar guiding active in PHD2, only the lock position is shown.

To make this change possible it was necessary to change the code for the internal guider, since the internal guider connected to the newImage event each time it started a capture, and deconnected when the image has been received. This is now changed such that this happens only once and the deconnect command is called only when capturing stops.

In addition, the phd2 JSON interface has been reworked and parameter arrays are replaced by proper JSON objects with key/value pairs to ensure correct PHD2 invocation.