Add developer option to not sync clock to real time when Ekos starts
Adds a checkbox in the Developer section of the KStars options menu to NOT sync to the real time clock when Ekos starts up. This is useful when trying recreate reported issues (i.e. run KStars at the same time as the report). It is intentionally done as a sort of double negative (check the box to not sync) as I wanted the normal case to be unchecked, like all the other developer options.