Focus Param Fix
A bug was reported by Hy where storing the Optical Train Combo in the Focus settings causes a "race condition" that results in settings not be saved correctly. There is code currently in place to prevent the Optical Train Combo being stored in settings so it's unclear why this happened.

This fix scrubs settings when they are first loaded to removed the Optical Train Combo if its present.