Bugfix creating a darks library using camera tabs
Creating a new dark library now works properly no matter from which camera tab it is used.

To make this work properly, it is no longer possible to change the optical train in the dark library UI. It always uses the optical train of the camera tab from which it is called.