Bug fix for optical train instance handling
Since the Scheduler is started before the devices are connected, it already has a connection to the optical train's instance. Therefore, when the manager calls start(), it must not release the static optical train's instance. Otherwise, the scheduler will not receive any updates when other modules change the optical train setup.