Fix timeDilation widget greyed out
time dilation widget is greyed out and cannot be activated. It is controlled by a checkable QGroupBox but this appears to be faulty.

Some sub widgets were set enabled and some disabled which seems to confuse the QGroupBox. Setting them all to enabled in designer seems to make it work correctly.