Focus incorrectly moves ROI to top left of image on non-star based focus measures
Bug reported in Focus whereby user was trying to use an ROI for solar imaging using the Sobel filter. Having selected the ROI and started Autofocus the ROI was moved (incorrectly) for the user selection to the top left of the image rendering ROI focus un-usable.

Not sure when this was broken. Fixed bug and tested on sims.