Focus: Terminate long running star detection when user aborts Autofocus
When the user aborts an Autofocus run, Focus currently check whether star detection is in progress. If so, it waits for it complete before resetting. Normally this is OK but if the reason the user is aborting Autofocus is that star extraction is running long then waiting for it to complete is problematic. Bug reported by Wolfgang.

This change waits 5 seconds then calls the StellarSolver abort() method for SEP star extraction.

Tested on sims. 
Wolfgang tested.