Draft: Focus history navigation
When focusing does not work as expected, it can be very helpful to navigate through the history of the focus frames to inspect the frames from all autofocus runs.

Therefore, a navigation panel is added to the Focus module. In order to save vertical space the display of detected stars, current HFR or FWHM is placed above the V-Curve, and no longer below. This saves an entire row on the right side.

The buttons are only active when no autofocus sequence is running.

Todos:
- [x] Replace the "Save Focus Images" option, define a limit of kept focus images to avoid storage overflow problems
- [x] Enable the navigation through all autofocus runs (as long as the images are present)
- [x] Change the navigation into an overlay of the focus view (similar to the overlay on the Manager tab),
- [x] Separate counters and navigation, optimize the right layout.
- [x] Display star measurements for the entire history

With the clear button, a single run can be deleted including the captured focus frames belonging to this run. The user will be asked whether the the captured frames should be moved to the trash, deleted or kept.