Optionally run Live Stacker as a separate process
Optionally run Live Stacker as a separate process. The benefit of this is that if it crashes it won't take KStars/Ekos with it.

If this option is selected, LiveStacker is started as a separate process...
![Screenshot_2025-07-20_at_16.19.34](/uploads/f575bbe2963a7b756ea7487c042b35bf/Screenshot_2025-07-20_at_16.19.34.png)

To run multiple Live Stackers as separate processes, select Live Stacker multiple times from the Sky Map.
To run multiple Live Stackers as tabs within 1 FitsViewer window which is run as a separate process, run the first Live Stacker from SkyMap to launch the separate process then run subsequent Live Stackers from the initial Live Stacker (FitsViewer) menu.

If the option isn't selected Live Stacker behaves as it does now.