Fixing some kstars file installs for the qt6 build for macOS and changing the phonon backend
The QT5 VLC and phonon backends we were using before no longer work in qt6.  This removes them from the qt6 build. It corrects an install location for notifications in qt6.  Also it fixes an annoying /kstars/fr install error that has appeared for years due to a missing environment variable.  It also sets the qt environment variable QT_MEDIA_BACKEND to Darwin so that it uses an existing backend that should work without VLC.