Remove OpenCV highgui component
The highgui component of OpenCV is useful for debugging but isn't strictly needed for production.

On Ubuntu including highgui doesn't cause any issues and it was included for some Focus methods.

With Live Stacker, more KStars users are trying openCV functions. One forum user running Mint ran into a problem whereby compiling with openCV worked fine but on startup KStars aborted with messages about GTK and Qt conflicting. The user was able to overcome the issue by compiling his own version of openCV and excluding GTK.

So the purpose of this MR is to remove the main cause of openCV possibly using GTK - highgui. This increases the chances of unix distros being compatible between KStars and OpenCV out of the box without requiring complicated builds.