Fixes observatory default weather source
Where multiple weather sources are present the current addWeatherSource implementation also overrides the Options::DefaultObservatoryWeatherSource with the first received weather source regardless of the previous saved default. It is currently only the order in which the weather devices are received in Ekos that determines the default.
The MR splits up the population of the weatherSourcesCombo list from the setting of the default source to correct the behaviour back to the intended, ie: default source always wins unless it doesn't exist in which case the first source wins.