Push-To Assistant
The new push-to assistant is intended as plate solving support for mechanical mounts in combination with a digital camera on a finder scope.

![image](/uploads/3c36c57b1eab3ce6bd1fddab6c0005ab/image.png){width=364 height=489}

**Setup**: Create an optical train with the Telescope Simulator as mount and configure your combination of finder scope and digital camera.

**Usage**
- start EKOS
- open the Push-to Assistant located in the Tools menu
- move your scope as good as possible to the target you want to find
- select the target from the catalog or enter its coordinates manually and press "Select target"
- press "Solve position" to determine the position your scope is currently pointing at
- as soon as the position has been solved, the assistant displays hints in which direction you should move your mount to be closer to the target
- correct your mount position and press "Solve position" again
- if you want to automatically repeat plate solving, configure the delay and press the "Repeat" button.

In order to make the assistant as easy as possible, EKOS can be started directly from the button in the status line at the bottom of the assistant:

![image](/uploads/f899b1ee1c546d402b3e4f6164a38119/image.png){width=190 height=35}

The LEDs on the right side show whether the Align and the Mount modules are currently active, which both are required for the assistant working properly.