Live Stacking Monitor
Live Stacking Monitor window is a popup from Live Stacker that shows a table of subs that match the chosen directory in Live Stacker.

![Screenshot_2025-10-01_at_05.09.57](/uploads/c2071361c1af689c1d6bc9e15a002996/Screenshot_2025-10-01_at_05.09.57.png)

The purpose is to allow analysis of Live Stacker, for example to allow investigation of bottlenecks in the stacking process.

When a sub is added to the watched directory, the sub is added to the Monitor's table. As the sub is processed by each step of the process information is updated in the table:
- Waiting to load. The sub is in the queue to be processed but Live Stacker is busy with other subs.
- Loading. The sub is loaded into memory.
- Plate solving. The sub is undergoing plate solving (if appropriate).
- Waiting to stack. The sub is waiting to be stacked (e.g. currently there are insufficient subs loaded to start a stack).
- Calibration. Dark / Flat calibration.
- Alignment.
- Stacking.

Table columns and sort order are configurable. Changing cells can be highlighted (or not).