Live Stacking - ImageMM Stacking Method
Implementation of the ImageMM stacking method as detailed here: https://iopscience.iop.org/article/10.3847/1538-3881/adfb72

ImageMM on the left, "regular" stack on the right...
![Screenshot_2025-10-09_at_22.47.49](/uploads/1d4b591536bcee25b15f02f1ea192936/Screenshot_2025-10-09_at_22.47.49.png)

This implementation strikes a balance between speed (it needs to be a Live Stacker) and fidelity.

To use: select ImageMM as the stacking method and play with the available controls.

This method is considerably more resource intensive than "regular" stacking because it uses an iterative approach.